added ytdl

+++ libs/sqlite_interface.py
@@ -5,12 +5,16 @@ import datetime
 from libs.sb_times_retrival import compare_sponsor_times
 
 
-class Db:
-    def __init__(self, db_name="yt_dl_sb.db", create_new=True):
+class Db(metaclass=obj.Singleton):
+    def __init__(self, db_name="yt_dl_sb.db", create_new=False):
         if create_new:
-            os.unlink(db_name)
+            if os.path.exists(db_name):
+                os.unlink(db_name)
         self.conn = sqlite3.connect(db_name)
         self.c = self.conn.cursor()
+        if create_new:
+            self.create_tables()
+
 
     def __enter__(self):
         return self
@@ -50,13 +54,23 @@ class Db:
 
     def update_last_fetched(self, ch: obj.Channel):
         # TODO: datetime to current timezone!
-        self.c.execute("update channels set last_fetched=? where channel_id=?",(str(datetime.datetime.now()),ch.channel_id))
+        self.c.execute("update channels set last_fetched=? where channel_id=?",
+                       (str(datetime.datetime.now()), ch.channel_id))
 
     def get_channel(self, channel_id: str):
         self.c.execute(f"SELECT id, channel_id, channel_name, last_fetched FROM channels WHERE channels.channel_id=?;",
                        (channel_id,))
         id, channel_id, channel_name, last_fetched = self.c.fetchone()
-        return obj.Channel(id, channel_id, channel_name, datetime.datetime.fromisoformat(last_fetched) if last_fetched else None)
+        return obj.Channel(id, channel_id, channel_name,
+                           datetime.datetime.fromisoformat(last_fetched) if last_fetched else None)
+
+    def get_all_channels(self):
+        self.c.execute("SELECT id, channel_id, channel_name, last_fetched FROM channels;")
+        return [
+            obj.Channel(id, channel_id, channel_name, datetime.datetime.fromisoformat(last_fetched) if last_fetched else None) for
+            id, channel_id, channel_name, last_fetched in self.c.fetchall()
+        ]
+
 
     def create_video(self, vid: obj.Video):
         self.c.execute('''insert into videos (video_id, status, downloaded_path) values (?,?,?)''',
@@ -70,17 +84,20 @@ class Db:
 
         return self.get_video(vid.video_id)
 
-
     def get_video(self, video_id: str):
-        self.c.execute(f"SELECT id, video_id, status, downloaded_path,latest_sponsortime_at FROM videos WHERE videos.video_id=?;", (video_id,))
-        id, video_id, status, downloaded_path,latest_sponsortime = self.c.fetchone()
-        self.c.execute(f"SELECT id, video_id, start, stop, votes FROM sponsor_times WHERE sponsor_times.video_id=?;", (video_id,))
+        self.c.execute(
+            f"SELECT id, video_id, status, downloaded_path,latest_sponsortime_at FROM videos WHERE videos.video_id=?;",
+            (video_id,))
+        id, video_id, status, downloaded_path, latest_sponsortime = self.c.fetchone()
+        self.c.execute(f"SELECT id, video_id, start, stop, votes FROM sponsor_times WHERE sponsor_times.video_id=?;",
+                       (video_id,))
         sponsor_times = [obj.SponsorTime(*st) for st in self.c.fetchall()]
 
-        return obj.Video(id, video_id, status, downloaded_path,sponsor_times,latest_sponsortime)
+        return obj.Video(id, video_id, status, downloaded_path, sponsor_times, latest_sponsortime)
 
     def update_video(self, vid: obj.Video):
-        self.c.execute(f"update videos set status=?, downloaded_path=? where video_id=?",(vid.status, vid.downloaded_path, vid.video_id))
+        self.c.execute(f"update videos set status=?, downloaded_path=? where video_id=?",
+                       (vid.status, vid.downloaded_path, vid.video_id))
         self.c.execute("delete from sponsor_times where video_id=?", (vid.video_id,))
         if vid.sponsor_times:
             sponsor_times_tuple = [(st.video_id, st.start, st.stop, st.votes) for st in vid.sponsor_times]
@@ -89,34 +106,33 @@ class Db:
         self.conn.commit()
         return self.get_video(video_id=vid.video_id)
 
-
-
     def update_video_sponsortimes_if_new(self, vid: obj.Video):
         vid_in_db = self.get_video(vid.video_id)
         if vid_in_db.sponsor_times is None and vid.sponsor_times is None:
             return False, vid
         if vid_in_db.sponsor_times is None and vid.sponsor_times:
             new_vid = self.update_video(vid)
-            return True,new_vid
+            return True, new_vid
         if vid_in_db.sponsor_times and vid.sponsor_times is None:
-            new_vid=self.update_video(vid)
-            return True,new_vid
+            new_vid = self.update_video(vid)
+            return True, new_vid
 
         for st in vid_in_db.sponsor_times:
             if not any([compare_sponsor_times(st, new_st) for new_st in vid.sponsor_times]):
-                new_vid=self.update_video(vid)
+                new_vid = self.update_video(vid)
                 return True, new_vid
         return False, vid
 
 
-
 if __name__ == "__main__":
-    db = Db()
-    db.create_tables()
-    db.create_channel(obj.Channel(None, "bla", "bla",None))
-    print(db.get_channel("bla"))
-    v = db.create_video(obj.Video(None, "123", "DOWNLOADING", None, [obj.SponsorTime(None, "123", "13", "12", 51), ],None))
-    db.update_video(obj.Video(None, "123", "DONE", "bla",[obj.SponsorTime(None, "123", "13", "123", 51),],None))
-    v.sponsor_times.append(obj.SponsorTime(None, "123", "13", "12", 51))
-    print(db.get_video("123"))
-    print(db.update_video_sponsortimes_if_new(v))
+    db = Db(create_new=True)
+
+    #db.create_tables()
+    #db.create_channel(obj.Channel(None, "bla", "bla", None))
+    #print(db.get_channel("bla"))
+    v = db.create_video(
+        obj.Video(None, "HeUietgDuVc", "FETCHED", None, [obj.SponsorTime(None, "123", "13", "12", 51), ], None))
+    #db.update_video(obj.Video(None, "123", "DONE", "bla", [obj.SponsorTime(None, "123", "13", "123", 51), ], None))
+    #v.sponsor_times.append(obj.SponsorTime(None, "123", "13", "12", 51))
+    #print(db.get_video("123"))
+    #print(db.update_video_sponsortimes_if_new(v))
